[Tests] Move MockConsole and IConsole to common folders
