+++ src/tests/Common/MockConsole.cs
@@ -1,12 +1,13 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
 using System.Diagnostics;
-using Microsoft.Diagnostics.Tools.Counters.Exporters;
+using System.IO;
+using Microsoft.Diagnostics.Tools.Common.Exporters;
 using Xunit;
 
-namespace DotnetCounters.UnitTests
+namespace Microsoft.Diagnostics.Tests.Common
 {
     internal class MockConsole : IConsole
     {
@@ -20,6 +21,7 @@ namespace DotnetCounters.UnitTests
             WindowHeight = height;
             Clear();
         }
+
         public int WindowHeight { get; init; }
 
         public int WindowWidth { get; init; }
