+++ src/tests/dotnet-counters/DotnetCounters.UnitTests.csproj
@@ -11,6 +11,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Include="..\Common\MockConsole.cs" Link="MockConsole.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="NewtonSoft.Json" Version="$(NewtonSoftJsonVersion)" />
   </ItemGroup>
